requirements.txt
requests
urllib3
python-dotenv
argparse